pyproject.toml
[project]
name = "aiengine"
version = "0.1.0"
description = "\"Find the songs/lyrics you love\""
authors = [
    {name = "y"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim (>=0.4.10,<0.5.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "pydantic-logfire (>=0.0.1,<0.0.2)",
    "logfire (>=4.0.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
